style: Move MediaQuery and friends to its own module.
Bug: 1468846
MozReview-Commit-ID: 3FRMnIHFwR3
